# backend/sample_data/openapi_auth.yaml
openapi: 3.0.3
info:
  title: PIH Auth API
  version: 1.0.0
  description: >
    API de registro e inicio de sesión del Portal Prestadores — Demo.
    Spec generado para vincular endpoints con las historias de Jira del
    proyecto PIH (login/registro), no es un contrato productivo real.
paths:
...
  /api/auth/login:
    post:
      operationId: iniciarSesion
      summary: Iniciar sesión
      tags: [Auth]
      description: >
        Autentica al usuario con email y contraseña. Implementa PIH-9
        (Inicio de sesión exitoso). Corrige PIH-16 (login no informa
        credenciales incorrectas) y PIH-18 (hotfix de pantalla congelada
        en login).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
                remember_me: { type: boolean }
      responses:
        "200":
          description: Login exitoso, devuelve token de sesión
        "401":
          description: Credenciales incorrectas
...